safari-534.53-branch/LayoutTests/platform/mac/accessibility/aria-tab-roles-expected.txt

Crust
Veges
Select Crust

This tests that the aria roles for tab, tabpanel and tablist work as expected correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tabList.role is 'AXRole: AXTabGroup'
PASS tab1.role is 'AXRole: AXRadioButton'
PASS tab1.title is 'AXTitle: Crust'
PASS tab1.childrenCount is 0
PASS tab2.role is 'AXRole: AXRadioButton'
PASS tab2.title is 'AXTitle: Veges'
PASS tabPanel.role is 'AXRole: AXGroup'
PASS tabPanel.subrole is 'AXSubrole: AXTabPanel'
PASS successfullyParsed is true

TEST COMPLETE
